+++ src/main/java/me/efekos/simpler/commands/SubCommand.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 efekos
+ * Copyright (c) 2024 efekos
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,10 +24,7 @@ package me.efekos.simpler.commands;
 
 import me.efekos.simpler.annotations.Command;
 import me.efekos.simpler.commands.syntax.Argument;
-import org.bukkit.Location;
-import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 import java.util.stream.Stream;