+++ CHANGELOG.md
@@ -101,4 +101,5 @@ keep running if the sender is a player.
 
 * **Added** `CommandNode.description` - Going to be used for auto-generated help menus.
 * **Added** `CommandNode.parent` - Going to be used for auto-generated help menus.
+* **Changed** License year to 2024.
 * **Deprecated** `ItemContent`, and it's needed classes. Everything here will be moved into [another repository](https://github.com/efekos/ItemContent).